requirements.txt
openai>=1.3.8
pandas>=1.5.0
scikit-learn>=1.3.0
numpy>=1.24.0
matplotlib>=3.6.0
seaborn>=0.12.0
jupyter>=1.0.0
pytest>=7.0.0
python-dotenv>=1.0.0
tqdm>=4.64.0
jsonlines>=3.1.0
pydantic>=2.0.0 